@@ -1,77 +1,82 @@
 {
-	"name": "roots/sage",
-	"type": "wordpress-theme",
-	"license": "MIT",
-	"description": "WordPress starter theme with a modern development workflow",
-	"homepage": "https://roots.io/sage/",
-	"authors": [
-		{
-			"name": "Ben Word",
-			"email": "ben@benword.com",
-			"homepage": "https://github.com/retlehs"
-		},
-		{
-			"name": "Scott Walkinshaw",
-			"email": "scott.walkinshaw@gmail.com",
-			"homepage": "https://github.com/swalkinshaw"
-		},
-		{
-			"name": "QWp6t",
-			"email": "hi@qwp6t.me",
-			"homepage": "https://github.com/qwp6t"
-		},
-		{
-			"name": "Brandon Nifong",
-			"email": "brandon@tendency.me",
-			"homepage": "https://github.com/log1x"
-		}
-	],
-	"keywords": [
-		"wordpress"
-	],
-	"support": {
-		"issues": "https://github.com/roots/sage/issues",
-		"forum": "https://discourse.roots.io/"
-	},
-	"autoload": {
-		"psr-4": {
-			"App\\": "app/"
-		}
-	},
-	"require": {
-		"php": "^7.4|^8.0",
-		"log1x/acf-composer": "^2.0",
-		"log1x/sage-directives": "^1.1",
-		"log1x/sage-svg": "^1.1",
-		"roots/acorn": "^2.0"
-	},
-	"require-dev": {
-		"squizlabs/php_codesniffer": "^3.6"
-	},
-	"suggest": {
-		"log1x/sage-directives": "A collection of useful Blade directives for WordPress and Sage (^1.0).",
-		"log1x/sage-svg": "A useful SVG directive for inlining SVG's within Blade views (^1.0)."
-	},
-	"config": {
-		"optimize-autoloader": true,
-		"preferred-install": "dist",
-		"sort-packages": true
-	},
-	"minimum-stability": "dev",
-	"prefer-stable": true,
-	"scripts": {
-		"lint": [
-			"phpcs --extensions=php --standard=PSR12 app"
-		],
-		"post-autoload-dump": [
-			"Roots\\Acorn\\ComposerScripts::postAutoloadDump"
-		]
-	},
-	"extra": {
-		"acorn": {
-			"providers": [
-				"App\\Providers\\ThemeServiceProvider"
-			]
-		}
-	}
+    "name": "roots/sage",
+    "type": "wordpress-theme",
+    "license": "MIT",
+    "description": "WordPress starter theme with a modern development workflow",
+    "homepage": "https://roots.io/sage/",
+    "authors": [
+        {
+            "name": "Ben Word",
+            "email": "ben@benword.com",
+            "homepage": "https://github.com/retlehs"
+        },
+        {
+            "name": "Scott Walkinshaw",
+            "email": "scott.walkinshaw@gmail.com",
+            "homepage": "https://github.com/swalkinshaw"
+        },
+        {
+            "name": "QWp6t",
+            "email": "hi@qwp6t.me",
+            "homepage": "https://github.com/qwp6t"
+        },
+        {
+            "name": "Brandon Nifong",
+            "email": "brandon@tendency.me",
+            "homepage": "https://github.com/log1x"
+        }
+    ],
+    "keywords": [
+        "wordpress"
+    ],
+    "support": {
+        "issues": "https://github.com/roots/sage/issues",
+        "forum": "https://discourse.roots.io/"
+    },
+    "autoload": {
+        "psr-4": {
+            "App\\": "app/"
+        }
+    },
+    "require": {
+        "php": "^8.0",
+        "log1x/acf-composer": "^2.0",
+        "log1x/sage-directives": "^1.1",
+        "log1x/sage-svg": "^1.1",
+        "roots/acorn": "^3.0"
+    },
+    "require-dev": {
+        "spatie/laravel-ignition": "^1.6",
+        "squizlabs/php_codesniffer": "3.7.1"
+    },
+    "suggest": {
+        "log1x/sage-directives": "A collection of useful Blade directives for WordPress and Sage (^1.0).",
+        "log1x/sage-svg": "A useful SVG directive for inlining SVG's within Blade views (^1.0)."
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": "dist",
+        "sort-packages": true,
+        "allow-plugins": {
+            "composer/installers": true,
+            "roots/wordpress-core-installer": true
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
+    "scripts": {
+        "lint": [
+            "phpcs --extensions=php --standard=PSR12 app"
+        ],
+        "post-autoload-dump": [
+            "Roots\\Acorn\\ComposerScripts::postAutoloadDump"
+        ]
+    },
+    "extra": {
+        "acorn": {
+            "providers": [
+                "App\\Providers\\ThemeServiceProvider"
+            ]
+        }
+    }
 }
