@@ -13,20 +13,19 @@ namespace App;
  */
 add_filter('excerpt_more', fn() => sprintf(' &hellip; <a href="%s">%s</a>', get_permalink(), __('Continued', 'sage')));
 
-if (class_exists('ACF')) {
-    // add_filter('acf/settings/show_admin', '__return_false');
-}
+// if (class_exists('ACF') && get_current_user_id() !== env('SUPER_ADMIN_ID')) {
+//     add_filter('acf/settings/show_admin', '__return_false');
+// }
 
 /**
- * Sets the location in the theme folder to save Field Group JSON files to
+ * This filter adds a "Sage Theme" Category for custom blocks,
+ * and sets it to the top most level so that it is more easily accessible
  */
-add_filter('acf/settings/save_json', fn() => get_template_directory() . '/acf-json');
-
-/**
- * Set the location in the theme folder to import Field Group JSON files from
- */
-add_filter('acf/settings/load_json', function ($paths) {
-    unset($paths[0]);
-    $paths[] = get_template_directory() . '/acf-json';
-    return $paths;
-});
+add_filter('block_categories_all', fn ($categories) => collect($categories)
+    ->prepend(
+        [
+            'slug'  => 'sage-theme',
+            'title' => 'Sage Theme Blocks',
+        ],
+    )
+    ->toArray());
