@@ -14,8 +14,9 @@ use function Roots\bundle;
  * @return void
  */
 add_action('wp_enqueue_scripts', function () {
-    // Uncomment to enable React on the site
-    // bundle('react')->enqueue();
+    // Uncomment to enable React or Vue on the site
+    // bundle('sageReact')->enqueue();
+    // bundle('vue')->enqueueJs();
     bundle('app')->enqueue();
 }, 100);
 
@@ -25,7 +26,21 @@ add_action('wp_enqueue_scripts', function () {
  * @return void
  */
 add_action('enqueue_block_editor_assets', function () {
+    // Uncomment to enable React or Vue on the site
+    // bundle('sageReact')->enqueue();
+    // bundle('vue')->enqueueJs();
     bundle('editor')->enqueue();
+}, 100);
+
+/**
+ * Registers an option CSS file and JS file for use within the wp-admin.
+ * Contains some "nice" styles for ACF, and there's an empty JS file.
+ * Uncomment to use
+ *
+ * @return void
+ */
+add_action('admin_enqueue_scripts', function () {
+    bundle('admin')->enqueue();
 }, 100);
 
 /**
@@ -54,11 +69,15 @@ add_action('after_setup_theme', function () {
 
     /**
      * Register the navigation menus.
+     *
      * @link https://developer.wordpress.org/reference/functions/register_nav_menus/
      */
-    register_nav_menus([
-        'primary_navigation' => __('Primary Navigation', 'sage'),
-    ]);
+    register_nav_menus(
+        [
+            'primary_navigation' => __('Primary Navigation', 'sage'),
+            'footer_navigation'  => __('Footer Navigation', 'sage'),
+        ]
+    );
 
     /**
      * Disable the default block patterns.
@@ -108,15 +127,15 @@ add_action('after_setup_theme', function () {
      * Updates and sets the WordPress image sizes
      */
     $options = [
-        'thumbnail_size_w' => 300,
-        'thumbnail_size_h' => 0,
-        'thumbnail_crop' => 0,
-        'medium_size_w' => 768,
-        'medium_size_h' => 0,
+        'thumbnail_size_w'    => 300,
+        'thumbnail_size_h'    => 0,
+        'thumbnail_crop'      => 0,
+        'medium_size_w'       => 768,
+        'medium_size_h'       => 0,
         'medium_large_size_w' => 1024,
         'medium_large_size_h' => 0,
-        'large_size_w' => 1280,
-        'large_size_h' => 0,
+        'large_size_w'        => 1280,
+        'large_size_h'        => 0,
     ];
 
     collect($options)
@@ -155,23 +174,6 @@ add_action('widgets_init', function () {
     ] + $config);
 });
 
-
-/**
- * Create options pages
- */
-if (function_exists('acf_add_options_page')) {
-    acf_add_options_page(
-        [
-            'page_title' => 'Theme General Settings',
-            'menu_title' => 'Theme Settings',
-            'menu_slug'  => 'theme-general-settings',
-            'capability' => 'edit_posts',
-            'redirect'   => false,
-            'post_id'    => 'theme_settings',
-        ]
-    );
-}
-
 /**
  * Disable Emoji's
  */
@@ -187,18 +189,12 @@ add_action('init', function () {
     /**
      * Filter out the tinymce emoji plugin.
      */
-    add_filter('tiny_mce_plugins', function ($plugins) {
-        if (is_array($plugins)) {
-            return array_diff($plugins, ['wpemoji']);
-        } else {
-            return [];
-        }
-    });
+
+    add_filter('tiny_mce_plugins', fn ($plugins) => is_array($plugins)
+        ? array_diff($plugins, ['wpemoji'])
+        : []);
 });
 
-/**
- * Registers and adds Admin scripts to the site
- */
-add_action('admin_enqueue_scripts', function () {
-    wp_enqueue_style('sage-admin-styles', get_template_directory_uri() . '/resources/admin_assets/css/admin_styles.min.css', [], null);
-});
+// Adds the ACF Nav menu custom field to the selectable options
+class_exists('ACF')
+    && add_action('acf/include_field_types', fn () => include_once 'Classes/AcfNavMenu.php');
