@@ -1,8 +1,9 @@
+// https://tailwindcss.com/docs/configuration
 module.exports = {
-    content: ["./index.php", "./app/**/*.php", "./resources/**/*.{php,vue,js}"],
+    content: ['./index.php', './app/**/*.php', './resources/**/*.{php,vue,js}'],
     theme: {
         extend: {
-            colors: {},
+            colors: {}, // Extend Tailwind's default colors
         },
     },
     plugins: [],
